DebouncedJob: Address incorrect Carbon object usage when persisting wait times
Previously when calling "addMinute()" on the minimum and maximum wait
Carbon objects directly, we were manipulating the actual minimum value
being cached.

This was exposed when the previous time bugs were fixed
and resulted in these DebouncedJobs always waiting for at least 1 minute + the
given minimum up to 1 minute + the given maximum.

To resolve this we just need to copy the Carbon object before adding the
minute for our cache TTL values