DebouncedJob: Add support for Job defined cache key generation
The previous and still currently default behavior of generating the
cache key for a debounced job by encoding the job itself was unique
enough but also would cause two identical jobs for the same parameters
to be considered different for things such as timestamps representing
updated_at/lastModified values.

This update allows individual Jobs to define how their cache key is
generated via a getDebounceCacheKey function.  For now this is
effectively undocumented as I wasn't sure if a Trait or interface would
be desirable and the default behavior is essentially unchanged